removes express and refactors code as a file to just be run in node. removes server.js. refactors start:dev and start:prod script. removes express as a dependency. recreates yarn.lock file.
