+++ journals/maiden-voyage.txt
@@ -1,0 +1,1 @@
+http://www.reddit.com/,https://www.reddit.com/user/WardCannon,https://www.reddit.com/r/IAmA/,https://www.reddit.com/subreddits/,https://www.reddit.com/r/tifu/
