+++ package.json
@@ -6,21 +6,23 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "npm run start:dev",
-    "start:production": "node server.js",
-    "start:dev": "nodemon -e js server.js"
+    "start:dev": "nodemon app.js",
+    "start:prod": "node app.js"
   },
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@github.com/illuminatedspace/passepartout.git"
+    "url": "http://git@github.com/illuminatedspace/passepartout.git"
   },
   "keywords": [
     "web",
     "scraper",
     "javascript",
     "node.js",
-    "fun"
+    "fun",
+    "promises",
+    "recursion"
   ],
-  "author": "Liz Phillips, lizkristinaphillips.com",
+  "author": "Liz Phillips",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/illuminatedspace/passepartout/issues"
@@ -28,7 +30,6 @@
   "homepage": "https://github.com/illuminatedspace/passepartout#readme",
   "dependencies": {
     "cheerio": "^0.22.0",
-    "express": "^4.15.2",
     "request": "^2.81.0"
   },
   "devDependencies": {
